Manually put ruby/jekyll in NSTask’s PATH.
On my machine, I have a jekyll binary in /usr/bin for some reason. On Ryan’s, it does not work because of this. This is a way for him to run it without spending too much time on this issue right now.

In the future, perhaps this can be auto-detected or something.